+++ js/photo.js
@@ -1,0 +1,81 @@
+
+// const getNumber = () => {
+//   let number = 0;
+
+
+//   return () => {
+//     number += 1;
+//     return number;
+
+//   };
+// };
+// const si = getNumber();
+// const count = getNumber();
+// const nextCount = getNumber();
+// // const flexCount = getNumber() - 0;
+
+// console.log(count());
+
+// let dog = 10;
+// let cat = 20;
+// const logToConsole = function(){
+
+//   dog = dog + 2;
+//   cat = cat + 10;
+//   console.log(dog, cat,);
+// };
+
+// logToConsole();
+
+// function makeCounter(count){
+//   return function(){
+//     return ++count;
+//   };
+
+// }
+// const counter = makeCounter(0);
+// const counter2 = makeCounter(0);
+
+
+// const doubleFive = (a,b) => {
+//   // или return 5 + 5
+//   const result = a + b;
+//   return result;
+// };
+// doubleFive(12,5);
+// console.log(doubleFive('12','a'));
+
+// // Определение
+// const run = () => {
+//   return 5;
+//   return 10;
+// };
+
+// // Что будет выведено на экран?
+// console.log(run());
+// const SETTING = {
+//   ARRAY_SIZE: 25;
+//   MIN_LIKES: 15;
+// }
+// const ARR = 'ARRAY_SIZE';
+
+// console.log(SETTING[ARR]);
+
+// const user = {
+//   name: 'Ivan Ivanov',
+//   age: 34,
+// };
+// console.log(user.name); // => 'Ivan Ivanov'
+
+// const user = {
+//   name: 60,
+//   age: 34,
+// };
+// const property = 'name';
+// console.log(user[property]); //
+
+// const hels = 'age';
+// console.log(user[hels]); //
+
+
+// console.log(property - hels);
